layout/base/tests/test_bug399951.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=399951
-->
<head>
  <title>Test for Bug 399951</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body style="direction: rtl" onload="document.body.style.direction = 'ltr';">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=399951">Mozilla Bug 399951</a>
<div style="white-space: pre;">
.i
  h
  f
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
 function test()
{
/** Test for Bug 399951 **/
  ok(true, "Should not crash");
  SimpleTest.finish();
}

  SimpleTest.requestFlakyTimeout("untriaged");
  setTimeout(test, 500);
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
